Use new custom tab bar in the root view.
